comment: 관련 없는 함수 분리, 주석 추가
